requirements.txt
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.4
peewee~=3.17.9
amadeus~=12.0.0
beautifulsoup4~=4.12.3

pycountry~=24.6.1
countryinfo~=0.1.2
python_telegram_bot_calendar~=1.0.5